p3_challenger

Struct SerializingChallenger64

pub struct SerializingChallenger64<F, Inner> { /* private fields */ }

Given a challenger that can observe and sample bytes, produces a challenger that is able to sample and observe field elements of a PrimeField64 field.

Observing:

Takes a field element will serialize it into a byte array and observe each byte.

Sampling:

Samples a field element in a prime field of size p by sampling uniformly an element in the range (0..1 << log_2(p)). This avoids modulo bias.

Implementations

impl<F: PrimeField64, Inner: CanObserve<u8>> SerializingChallenger64<F, Inner>

pub const fn new(inner: Inner) -> Self

impl<F, H> SerializingChallenger64<F, HashChallenger<u8, H, 32>>
where F: PrimeField64, H: CryptographicHasher<u8, [u8; 32]>,

pub const fn from_hasher(initial_state: Vec<u8>, hasher: H) -> Self